Package linux artifacts into .deb files
Problem: A customer requested that we bundle linux assets into .deb
files to make them easier to install.

Solution: Extend the GitHub Actions workflow to package a deb for
aarch64 and amd64 linux.

Testing: The PR for this commit will run the workflow, which can be used
to verify success.