Relax GLIBC Requirements for Linux Artifacts
Problem: We were using PyInstaller to package standalone executables
of `roboto` and `roboto-agent` for Linux + Windows. This bundles a
CPython interpreter based on the environment of the machine running
PyInstaller. Because the GitHub actions workers all use relatively new
versions of Ubuntu (20 earliest for x86, 22 earliest for arm), this
included a requirement on a pretty recent version of GLIBC, which
caused compatibility issues when try to run the command on an older
linux distro.

Solution: Switch to building scie executables for linux/windows, which
was what we were doing for macos anyway. These bundle super portable
Python interpreters, with the linux one only requiring GLIBC 2.17

Testing: Ran in CI, then ran in a docker container for an Ubuntu 14
based Arm environment.